refactor(core): iconify sources as two standalone linear implementations
Dissolves the createIconifySource skeleton and its IconifyBackend/
IconifySourceSpec/createPackAnchor abstractions: they existed to force the
local and API sources through one factory, sharing ~30 lines of real logic
at the price of ~120 lines of indirection (including a loadPack(names)
parameter one implementation ignored).

Each source now reads top-to-bottom as one linear closure, the way
localSource does - the pack anchor becomes two closure variables. What's
genuinely shared (allowlist dedupe/partition/rejection, pack-to-entries
resolution) is four small named helpers in iconify/shared.ts, shared by
function call rather than by framework.
