+++ packages/core/src/content/live.ts
@@ -2,7 +2,8 @@ export { createLiveIconLoader } from "./liveLoader.js";
 export type { LiveIconLoaderOptions } from "./liveLoader.js";
 export { liveIconCollections } from "./liveCollections.js";
 export type { LiveIconCollectionConfig } from "./liveCollections.js";
-export { iconifyLocalSource, iconifyApiSource } from "./iconify/source.js";
+export { iconifyLocalSource } from "./iconify/localSource.js";
+export { iconifyApiSource } from "./iconify/apiSource.js";
 export { mergeSources } from "./compositeSource.js";
 export type { CompositeSource } from "./compositeSource.js";
 export { AstroIconError } from "../internal/error.js";
